+++ SPD5/data.txt
@@ -1,21 +1,19 @@
-20  3
-  55   28  352
- 248   72  460
- 167   62  491
- 505   32  276
- 336    4  757
- 119   75  539
- 393   63  671
- 219   32   12
-  33   34  496
- 194    4  449
- 485   87  570
- 100   60  381
-  21   96  390
-  98   17  346
- 442   51  681
-  52   68   35
- 223   18  684
-  54   87  576
- 421    2  692
- 508   28  119
+18  3
+   5   15   35
+ 481   41  426
+ 473   62   49
+ 278   92  408
+  33   78  227
+ 164   64  662
+ 572   18   67
+ 554   59    1
+ 536   37  125
+ 552   56  501
+ 124   87  437
+  86   84  494
+ 534   88  495
+  85   68  643
+ 653   72  220
+ 157   74  460
+ 174   68  220
+ 309   10  362
